Temp: shake 임시저장
